requirements.txt
flask==3.0.0
langchain==0.1.0
wikipedia==1.4.0
duckduckgo-search==4.1.1
google-search-results==2.4.2
python-dotenv==1.0.0
openai==0.28.1 